refactor(#40): modify chat rooms dto naming
