+++ src/chat/swagger-decorators/get-chat-rooms.decorator.ts
@@ -6,7 +6,7 @@ import {
   ApiResponse,
   getSchemaPath,
 } from '@nestjs/swagger';
-import { ChatRoomsDto } from '../dto/chat-rooms.dto';
+import { ChatRoomsDto } from '../dto/chat-room.dto';
 
 export function ApiGetChatRooms() {
   return applyDecorators(
